New Efficency

+++ dtp/dtpHost.h
@@ -57,7 +57,7 @@ class dtpHost : public FIFONode {
     void set_normal_cookie();
     bool check_inorder_packets();
     void send_immediately(bool syn_temp,bool ack_temp,bool fin_temp,unsigned int id_temp,Time time_temp,bool);
-    int  check_congestion(Packet* pkt_p);
+    unsigned short  check_congestion(Packet* pkt_p);
     void delete_timer_cookie(unsigned int pack_no);
